plugins/moondream2/requirements.txt
einops
Pillow
